requirements.txt
python-telegram-bot
speedtest-cli==2.1.3